@@ -2,7 +2,9 @@ class Purchase < ActiveRecord::Base
   default_scope { order(created_at: :desc) }
 
   belongs_to  :user
-  has_many    :items, dependent: :destroy
-  validates_presence_of :items, :user
+  has_many	  :itemizables, dependent: :destroy
+  has_many    :items, through: :itemizables
+
+  validates_presence_of :itemizables, :user
 
 end
