@@ -9,10 +9,12 @@ describe EmailProcessor do
   before(:each) do
     Delayed::Worker.delay_jobs = false
     @email = FactoryGirl.build(:email, to: ["#{@user.endpoint_email}@#{EMAIL_URI}"])
+    @email.to = {email: @email.to}
+    @email.from = {email: @email.from}
   end
 
   it "should send right mailer if user is not found" do
-    @email.stubs(:to).returns(["non-existant-email@#{EMAIL_URI}"])
+    @email.stubs(:to).returns({email: ["non-existant-email@#{EMAIL_URI}"]})
     UserMailer.any_instance.expects(:cannot_find_account).with(@email, 'non-existant-email')
 
     EmailProcessor.process(@email)
